Feature/fight

+++ src/main/java/com/teamobi/mobiarmy2/constant/SpinWheelConstants.java
@@ -4,7 +4,7 @@ public class SpinWheelConstants {
     public static final int XU_COST = 1000; // Cost in Xu for spinning the wheel
     public static final int LUONG_COST = 1; // Cost in Luong for spinning the wheel
 
-    public static final int[] ITEM_COUNTS = {1, 5, 10, 15};
+    public static final byte[] ITEM_COUNTS = {1, 5, 10, 15};
     public static final int[] ITEM_PROBABILITIES = {400, 300, 200, 100};
 
     public static final int[] XU_COUNTS = {500, 1000, 5000, 10000};
